Improve loadtxt path handling
path can be path to directory of a summary file or complete path to the summary file. If a directory is specified and no summary is present in the directory, collect_logs will be called to generate summary file. If an incorrect summary filename is specified, the user is alerted of any summary files in the directory of the wrong filename, or that log files are present.