Implement parity filter for level_plot
This was actually a bug in the program, that the parity of the level plot was taken as the parity of the ground state, but the function did not have any actual filtering of parity meaning that both + and - were counted as if all were of the ground state parity. Now, the levels are actually filtered based on parity. If no filter is given, then the ground state parity is chosen. The next step is to include both parities in the plot by default, while keeping the possibility of filtering.