Change level_plot behaviour
Subtracting the ground state energy is no longer automatically done in the general_utilities level_plot. Instead, the Compare class does not use relative energy by default as it can be misleading, while the kshell_utilities level_plot does use the relative energy as default but it can be toggled.