Change B dist normalisation and PT func return values
The B distributions are now normalised in a more robust way by dividing the distributions by the integral of the distributions so that the normalised distributions integrate to 1. I have also chosen to not use the 0th bin in the plots and rather start at the 1st bin because of the exponential behaviour of the PT distribution as x approaches zero.